move formatNumber into shared
