+++ viz-lib/src/visualizations/custom/shared/formatNumber.ts
